+++ microservice-transaction-sample/web/src/services/requests/create-post.ts
@@ -1,7 +1,6 @@
 "use server";
 
 import { $ } from "zx";
-import { getUserId } from "./authentication-util";
 
 export const createPost = async (
   userId: number,
