+++ microservice-transaction-sample/web/src/services/requests/type.ts
@@ -1,10 +1,5 @@
 "use server";
 
-export const getUserId = async () => {
-  // TODO: implement this function
-  return 1;
-};
-
 export interface UserResponse {
   id: number;
   name: string;
